change in question and question-asnwer enpoint

+++ src/main/java/cu/edu/mes/sigenu/training/service/QuestionServiceImpl.java
@@ -2,6 +2,11 @@ package cu.edu.mes.sigenu.training.service;
 
 import java.util.List;
 
+import cu.edu.mes.sigenu.training.core.model.QuestionAnswer;
+import cu.edu.mes.sigenu.training.core.model.QuestionCarrer;
+import cu.edu.mes.sigenu.training.core.service.QuestionAnswerService;
+import cu.edu.mes.sigenu.training.core.service.QuestionCarrerService;
+import org.modelmapper.internal.bytebuddy.implementation.bytecode.Throw;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -14,6 +19,12 @@ public class QuestionServiceImpl implements QuestionService {
 
 	@Autowired
 	private QuestionRepository questionRepository;
+
+	@Autowired
+    private QuestionAnswerService questionAnswerService;
+
+	@Autowired
+    private QuestionCarrerService questionCarrerService;
 	
 	@Override
 	public List<Question> listAll() {
@@ -49,13 +60,26 @@ public class QuestionServiceImpl implements QuestionService {
 	public Question update(Question item) {
 		Question itemInDb = findById(item.getId());
 		itemInDb.setQuestion(item.getQuestion());
-		
+		itemInDb.setGroupQuestionId(item.getGroupQuestionId());
 		return questionRepository.save(itemInDb);
 	}
 
 	@Override
-	public void delete(Integer id) {
-		questionRepository.deleteById(id);
+	public void delete(Integer id) throws Exception {
+        Question question = findById(id);
+        if(question.getQuestionnaireList().size() != 0){
+            throw new Exception();
+        }
+        for(QuestionAnswer questionAnswer:question.getQuestionAnswerList()){
+            if(questionAnswer.getStudentAnswerList().size()!=0){
+                throw new Exception();
+            }
+        }
+        for(QuestionAnswer questionAnswer:question.getQuestionAnswerList()){
+            questionAnswerService.delete(questionAnswer.getId());
+        }
+        questionCarrerService.delete(question.getQuestionCarrerList().get(0).getId());
+        questionRepository.deleteById(id);
 	}
